added query for interact api
