+++ src/API/Query/query.js
@@ -1,5 +1,5 @@
-import { useQuery,useQueryClient } from "@tanstack/react-query"
-import {  getGraphData,getAIrecommendationsData, getPerformanceMatrics, getMaintainceData, getGraphsData, getProjectData} from "../APICalls/api"
+import { useQuery,useQueryClient,useMutation } from "@tanstack/react-query"
+import {  getGraphData,getAIrecommendationsData, getPerformanceMatrics, getMaintainceData, getGraphsData, getProjectData, getChatBotResponse} from "../APICalls/api"
 
 export const useGetFetchQuery = (key) => {
     const queryClient = useQueryClient();
@@ -55,3 +55,10 @@ export const useGetProjectData=()=>{
         queryFn:()=>getProjectData()
     })
 }
+
+
+export const useChatBotMutation = () => {
+  return useMutation({
+    mutationFn: getChatBotResponse,
+  });
+};
